requirements.txt
Django==3.2.7
psycopg2-binary>=2.8
djangorestframework
